+++ node/tests/ServerModules.test.ts
@@ -49,16 +49,16 @@ describe("Server Module Tests", () => {
         await cluster.close();
     }, TIMEOUT);
 
-    describe("GlideJson", () => {
-        let client: GlideClusterClient;
+    describe.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        "GlideJson",
+        (protocol) => {
+            let client: GlideClusterClient;
 
-        afterEach(async () => {
-            await flushAndCloseClient(true, cluster.getAddresses(), client);
-        });
+            afterEach(async () => {
+                await flushAndCloseClient(true, cluster.getAddresses(), client);
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "check modules loaded",
-            async (protocol) => {
+            it("check modules loaded", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -71,12 +71,9 @@ describe("Server Module Tests", () => {
                 });
                 expect(info).toContain("# json_core_metrics");
                 expect(info).toContain("# search_index_stats");
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.set and json.get tests",
-            async (protocol) => {
+            it("json.set and json.get tests", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -97,12 +94,12 @@ describe("Server Module Tests", () => {
                 ).toBe("OK");
 
                 // JSON.get
-                let result = await GlideJson.get(client, key, { paths: ["."] });
+                let result = await GlideJson.get(client, key, { path: "." });
                 expect(JSON.parse(result.toString())).toEqual(jsonValue);
 
                 // JSON.get with array of paths
                 result = await GlideJson.get(client, key, {
-                    paths: ["$.a", "$.b"],
+                    path: ["$.a", "$.b"],
                 });
                 expect(JSON.parse(result.toString())).toEqual({
                     "$.a": [1.0],
@@ -112,19 +109,16 @@ describe("Server Module Tests", () => {
                 // JSON.get with non-existing key
                 expect(
                     await GlideJson.get(client, "non_existing_key", {
-                        paths: ["$"],
+                        path: ["$"],
                     }),
                 );
 
                 // JSON.get with non-existing path
-                result = await GlideJson.get(client, key, { paths: ["$.d"] });
+                result = await GlideJson.get(client, key, { path: "$.d" });
                 expect(result).toEqual("[]");
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.set and json.get tests with multiple value",
-            async (protocol) => {
+            it("json.set and json.get tests with multiple value", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -149,7 +143,7 @@ describe("Server Module Tests", () => {
 
                 // JSON.get with deep path
                 let result = await GlideJson.get(client, key, {
-                    paths: ["$..c"],
+                    path: "$..c",
                 });
                 expect(JSON.parse(result.toString())).toEqual([true, 1, 2]);
 
@@ -159,18 +153,15 @@ describe("Server Module Tests", () => {
                 ).toBe("OK");
 
                 // verify JSON.set result
-                result = await GlideJson.get(client, key, { paths: ["$..c"] });
+                result = await GlideJson.get(client, key, { path: "$..c" });
                 expect(JSON.parse(result.toString())).toEqual([
                     "new_value",
                     "new_value",
                     "new_value",
                 ]);
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.set conditional set",
-            async (protocol) => {
+            it("json.set conditional set", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -200,7 +191,7 @@ describe("Server Module Tests", () => {
                     }),
                 ).toBeNull();
                 let result = await GlideJson.get(client, key, {
-                    paths: [".a"],
+                    path: ".a",
                 });
                 expect(result).toEqual("1");
 
@@ -209,14 +200,11 @@ describe("Server Module Tests", () => {
                         conditionalChange: ConditionalChange.ONLY_IF_EXISTS,
                     }),
                 ).toBe("OK");
-                result = await GlideJson.get(client, key, { paths: [".a"] });
+                result = await GlideJson.get(client, key, { path: ".a" });
                 expect(result).toEqual("4.5");
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.get formatting",
-            async (protocol) => {
+            it("json.get formatting", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -235,7 +223,7 @@ describe("Server Module Tests", () => {
                 ).toBe("OK");
                 // JSON.get with formatting options
                 let result = await GlideJson.get(client, key, {
-                    paths: ["$"],
+                    path: "$",
                     indent: "  ",
                     newline: "\n",
                     space: " ",
@@ -246,7 +234,7 @@ describe("Server Module Tests", () => {
                 expect(result).toEqual(expectedResult1);
                 // JSON.get with different formatting options
                 result = await GlideJson.get(client, key, {
-                    paths: ["$"],
+                    path: "$",
                     indent: "~",
                     newline: "\n",
                     space: "*",
@@ -255,12 +243,9 @@ describe("Server Module Tests", () => {
                 const expectedResult2 =
                     '[\n~{\n~~"a":*1,\n~~"b":*2,\n~~"c":*{\n~~~"d":*3,\n~~~"e":*4\n~~}\n~}\n]';
                 expect(result).toEqual(expectedResult2);
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.arrinsert",
-            async (protocol) => {
+            it("json.arrinsert", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -364,12 +349,9 @@ describe("Server Module Tests", () => {
                 expect(
                     JSON.parse((await GlideJson.get(client, key)) as string),
                 ).toEqual(expected);
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.arrpop",
-            async (protocol) => {
+            it("json.arrpop", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -401,7 +383,7 @@ describe("Server Module Tests", () => {
 
                 // Even if only one array element was returned, ensure second array at `..a` was popped
                 doc = (await GlideJson.get(client, key, {
-                    paths: ["$..a"],
+                    path: ["$..a"],
                 })) as string;
                 expect(doc).toEqual("[[],[3,4],42]");
 
@@ -417,12 +399,9 @@ describe("Server Module Tests", () => {
                     "OK",
                 );
                 expect(await GlideJson.arrpop(client, key)).toEqual("42");
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.arrlen",
-            async (protocol) => {
+            it("json.arrlen", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -458,12 +437,9 @@ describe("Server Module Tests", () => {
                     await GlideJson.set(client, key, "$", "[1, 2, 3, 4]"),
                 ).toBe("OK");
                 expect(await GlideJson.arrlen(client, key)).toEqual(4);
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.toggle tests",
-            async (protocol) => {
+            it("json.toggle tests", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -516,12 +492,9 @@ describe("Server Module Tests", () => {
                 await expect(
                     GlideJson.toggle(client, "non_existing_key", { path: "$" }),
                 ).rejects.toThrow(RequestError);
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.del tests",
-            async (protocol) => {
+            it("json.del tests", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -548,13 +521,13 @@ describe("Server Module Tests", () => {
                     await GlideJson.del(client, key, { path: "..path" }),
                 ).toBe(0);
 
-                // deleting existing paths
+                // deleting existing path
                 expect(await GlideJson.del(client, key, { path: "$..a" })).toBe(
                     2,
                 );
-                expect(
-                    await GlideJson.get(client, key, { paths: ["$..a"] }),
-                ).toBe("[]");
+                expect(await GlideJson.get(client, key, { path: "$..a" })).toBe(
+                    "[]",
+                );
                 expect(
                     await GlideJson.set(
                         client,
@@ -567,12 +540,12 @@ describe("Server Module Tests", () => {
                     2,
                 );
                 await expect(
-                    GlideJson.get(client, key, { paths: ["..a"] }),
+                    GlideJson.get(client, key, { path: "..a" }),
                 ).rejects.toThrow(RequestError);
 
                 // verify result
                 const result = await GlideJson.get(client, key, {
-                    paths: ["$"],
+                    path: "$",
                 });
                 expect(JSON.parse(result as string)).toEqual([
                     { b: { b: 2.5, c: true } },
@@ -604,7 +577,7 @@ describe("Server Module Tests", () => {
                 expect(await GlideJson.del(client, key)).toBe(1);
                 expect(await GlideJson.del(client, key)).toBe(0);
                 expect(
-                    await GlideJson.get(client, key, { paths: ["$"] }),
+                    await GlideJson.get(client, key, { path: "$" }),
                 ).toBeNull();
 
                 // non-existing keys
@@ -618,12 +591,9 @@ describe("Server Module Tests", () => {
                         path: ".",
                     }),
                 ).toBe(0);
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.forget tests",
-            async (protocol) => {
+            it("json.forget tests", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -654,9 +624,9 @@ describe("Server Module Tests", () => {
                 expect(
                     await GlideJson.forget(client, key, { path: "$..a" }),
                 ).toBe(2);
-                expect(
-                    await GlideJson.get(client, key, { paths: ["$..a"] }),
-                ).toBe("[]");
+                expect(await GlideJson.get(client, key, { path: "$..a" })).toBe(
+                    "[]",
+                );
                 expect(
                     await GlideJson.set(
                         client,
@@ -669,12 +639,12 @@ describe("Server Module Tests", () => {
                     await GlideJson.forget(client, key, { path: "..a" }),
                 ).toBe(2);
                 await expect(
-                    GlideJson.get(client, key, { paths: ["..a"] }),
+                    GlideJson.get(client, key, { path: "..a" }),
                 ).rejects.toThrow(RequestError);
 
                 // verify result
                 const result = await GlideJson.get(client, key, {
-                    paths: ["$"],
+                    path: "$",
                 });
                 expect(JSON.parse(result as string)).toEqual([
                     { b: { b: 2.5, c: true } },
@@ -710,7 +680,7 @@ describe("Server Module Tests", () => {
                 expect(await GlideJson.forget(client, key)).toBe(1);
                 expect(await GlideJson.forget(client, key)).toBe(0);
                 expect(
-                    await GlideJson.get(client, key, { paths: ["$"] }),
+                    await GlideJson.get(client, key, { path: "$" }),
                 ).toBeNull();
 
                 // non-existing keys
@@ -724,12 +694,9 @@ describe("Server Module Tests", () => {
                         path: ".",
                     }),
                 ).toBe(0);
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.type tests",
-            async (protocol) => {
+            it("json.type tests", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -792,12 +759,9 @@ describe("Server Module Tests", () => {
                 expect(
                     await GlideJson.type(client, "non_existing", { path: "." }),
                 ).toBeNull();
-            },
-        );
+            });
 
-        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
-            "json.resp tests",
-            async (protocol) => {
+            it("json.resp tests", async () => {
                 client = await GlideClusterClient.createClient(
                     getClientConfigurationOption(
                         cluster.getAddresses(),
@@ -921,9 +885,183 @@ describe("Server Module Tests", () => {
                 expect(
                     await GlideJson.resp(client, "nonexistent_key"),
                 ).toBeNull();
-            },
-        );
-    });
+            });
+
+            it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+                "json.strlen tests",
+                async (protocol) => {
+                    client = await GlideClusterClient.createClient(
+                        getClientConfigurationOption(
+                            cluster.getAddresses(),
+                            protocol,
+                        ),
+                    );
+                    const key = uuidv4();
+                    const jsonValue = {
+                        a: "foo",
+                        nested: { a: "hello" },
+                        nested2: { a: 31 },
+                    };
+                    // setup
+                    expect(
+                        await GlideJson.set(
+                            client,
+                            key,
+                            "$",
+                            JSON.stringify(jsonValue),
+                        ),
+                    ).toBe("OK");
+
+                    expect(
+                        await GlideJson.strlen(client, key, { path: "$..a" }),
+                    ).toEqual([3, 5, null]);
+                    expect(
+                        await GlideJson.strlen(client, key, { path: "a" }),
+                    ).toBe(3);
+
+                    expect(
+                        await GlideJson.strlen(client, key, {
+                            path: "$.nested",
+                        }),
+                    ).toEqual([null]);
+                    expect(
+                        await GlideJson.strlen(client, key, { path: "$..a" }),
+                    ).toEqual([3, 5, null]);
+
+                    expect(
+                        await GlideJson.strlen(client, "non_existing_key", {
+                            path: ".",
+                        }),
+                    ).toBeNull();
+                    expect(
+                        await GlideJson.strlen(client, "non_existing_key", {
+                            path: "$",
+                        }),
+                    ).toBeNull();
+                    expect(
+                        await GlideJson.strlen(client, key, {
+                            path: "$.non_existing_path",
+                        }),
+                    ).toEqual([]);
+
+                    // error case
+                    await expect(
+                        GlideJson.strlen(client, key, { path: "nested" }),
+                    ).rejects.toThrow(RequestError);
+                    await expect(GlideJson.strlen(client, key)).rejects.toThrow(
+                        RequestError,
+                    );
+                },
+            );
+
+            it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+                "json.strappend tests",
+                async (protocol) => {
+                    client = await GlideClusterClient.createClient(
+                        getClientConfigurationOption(
+                            cluster.getAddresses(),
+                            protocol,
+                        ),
+                    );
+                    const key = uuidv4();
+                    const jsonValue = {
+                        a: "foo",
+                        nested: { a: "hello" },
+                        nested2: { a: 31 },
+                    };
+                    // setup
+                    expect(
+                        await GlideJson.set(
+                            client,
+                            key,
+                            "$",
+                            JSON.stringify(jsonValue),
+                        ),
+                    ).toBe("OK");
+
+                    expect(
+                        await GlideJson.strappend(client, key, '"bar"', {
+                            path: "$..a",
+                        }),
+                    ).toEqual([6, 8, null]);
+                    expect(
+                        await GlideJson.strappend(
+                            client,
+                            key,
+                            JSON.stringify("foo"),
+                            {
+                                path: "a",
+                            },
+                        ),
+                    ).toBe(9);
+
+                    expect(
+                        await GlideJson.get(client, key, { path: "." }),
+                    ).toEqual(
+                        JSON.stringify({
+                            a: "foobarfoo",
+                            nested: { a: "hellobar" },
+                            nested2: { a: 31 },
+                        }),
+                    );
+
+                    expect(
+                        await GlideJson.strappend(
+                            client,
+                            key,
+                            JSON.stringify("bar"),
+                            {
+                                path: "$.nested",
+                            },
+                        ),
+                    ).toEqual([null]);
+
+                    await expect(
+                        GlideJson.strappend(
+                            client,
+                            key,
+                            JSON.stringify("bar"),
+                            {
+                                path: ".nested",
+                            },
+                        ),
+                    ).rejects.toThrow(RequestError);
+                    await expect(
+                        GlideJson.strappend(client, key, JSON.stringify("bar")),
+                    ).rejects.toThrow(RequestError);
+
+                    expect(
+                        await GlideJson.strappend(
+                            client,
+                            key,
+                            JSON.stringify("try"),
+                            {
+                                path: "$.non_existing_path",
+                            },
+                        ),
+                    ).toEqual([]);
+
+                    await expect(
+                        GlideJson.strappend(
+                            client,
+                            key,
+                            JSON.stringify("try"),
+                            {
+                                path: ".non_existing_path",
+                            },
+                        ),
+                    ).rejects.toThrow(RequestError);
+                    await expect(
+                        GlideJson.strappend(
+                            client,
+                            "non_existing_key",
+                            JSON.stringify("try"),
+                        ),
+                    ).rejects.toThrow(RequestError);
+                },
+            );
+        },
+    );
 
     describe("GlideFt", () => {
         let client: GlideClusterClient;
@@ -1144,5 +1282,78 @@ describe("Server Module Tests", () => {
                 expect((e as Error).message).toContain("Index does not exist");
             }
         });
+
+        it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+            "FT.INFO ft.info",
+            async (protocol) => {
+                client = await GlideClusterClient.createClient(
+                    getClientConfigurationOption(
+                        cluster.getAddresses(),
+                        protocol,
+                    ),
+                );
+
+                const index = uuidv4();
+                expect(
+                    await GlideFt.create(
+                        client,
+                        Buffer.from(index),
+                        [
+                            {
+                                type: "VECTOR",
+                                name: "$.vec",
+                                alias: "VEC",
+                                attributes: {
+                                    algorithm: "HNSW",
+                                    distanceMetric: "COSINE",
+                                    dimensions: 42,
+                                },
+                            },
+                            { type: "TEXT", name: "$.name" },
+                        ],
+                        { dataType: "JSON", prefixes: ["123"] },
+                    ),
+                ).toEqual("OK");
+
+                let response = await GlideFt.info(client, Buffer.from(index));
+
+                expect(response).toMatchObject({
+                    index_name: index,
+                    key_type: "JSON",
+                    key_prefixes: ["123"],
+                    fields: [
+                        {
+                            identifier: "$.name",
+                            type: "TEXT",
+                            field_name: "$.name",
+                            option: "",
+                        },
+                        {
+                            identifier: "$.vec",
+                            type: "VECTOR",
+                            field_name: "VEC",
+                            option: "",
+                            vector_params: {
+                                distance_metric: "COSINE",
+                                dimension: 42,
+                            },
+                        },
+                    ],
+                });
+
+                response = await GlideFt.info(client, index, {
+                    decoder: Decoder.Bytes,
+                });
+                expect(response).toMatchObject({
+                    index_name: Buffer.from(index),
+                });
+
+                expect(await GlideFt.dropindex(client, index)).toEqual("OK");
+                // querying a missing index
+                await expect(GlideFt.info(client, index)).rejects.toThrow(
+                    "Index not found",
+                );
+            },
+        );
     });
 });